Improve build times by cross compiling ARM flavors on Linux

Related work items: #2302930

+++ ci/run-tests.sh
@@ -157,29 +157,28 @@ perl $SCRIPT_DIR/generateTestSettingsJson.pl \
     --LongRunning "false"
 
 # For Python, these entries need to be in the root
-cp -rf tests/input .
-cp $binaryDir/test.settings.json .
-cp ./tests/test.defaults.json .
-cp ./tests/test.audio.utterances.json .
-cp ./tests/test.subscriptions.regions.json .
+[[ ! -d ./input ]] && cp -rf tests/input .
+[[ ! -e ./test.defaults.json ]] && cp ./tests/test.defaults.json .
+[[ ! -e ./test.audio.utterances.json ]] && cp ./tests/test.audio.utterances.json .
+[[ ! -e ./test.subscriptions.regions.json ]] && cp ./tests/test.subscriptions.regions.json .
 
-[[ -d $binaryDir/input ]] && cp -rf tests/input $binaryDir
-[[ -e $binaryDir/test.defaults.json ]] && cp ./tests/test.defaults.json $binaryDir
-[[ -e $binaryDir/test.audio.utterances.json ]] && cp ./tests/test.audio.utterances.json $binaryDir
-[[ -e $binaryDir/test.subscriptions.regions.json ]] && cp ./tests/test.subscriptions.regions.json $binaryDir
+[[ ! -d $binaryDir/input ]] && cp -rf tests/input $binaryDir
+[[ ! -e $binaryDir/test.defaults.json ]] && cp ./tests/test.defaults.json $binaryDir
+[[ ! -e $binaryDir/test.audio.utterances.json ]] && cp ./tests/test.audio.utterances.json $binaryDir
+[[ ! -e $binaryDir/test.subscriptions.regions.json ]] && cp ./tests/test.subscriptions.regions.json $binaryDir
 
 if [ -d buildfromdropx64 ]; then
-    cp -rf tests/input buildfromdropx64/bin
-    cp ./tests/test.defaults.json buildfromdropx64/bin
-    cp ./tests/test.audio.utterances.json buildfromdropx64/bin
-    cp ./tests/test.subscriptions.regions.json buildfromdropx64/bin
+    [[ ! -d buildfromdropx64/bin/input ]] && cp -rf tests/input buildfromdropx64/bin
+    [[ ! -e buildfromdropx64/bin/test.defaults.json ]] && cp ./tests/test.defaults.json buildfromdropx64/bin
+    [[ ! -e buildfromdropx64/bin/test.audio.utterances.json ]] && cp ./tests/test.audio.utterances.json buildfromdropx64/bin
+    [[ ! -e buildfromdropx64/bin/test.subscriptions.regions.json ]] && cp ./tests/test.subscriptions.regions.json buildfromdropx64/bin
 fi
 
 if [ -d buildfromdropx86 ]; then
-    cp -rf tests/input buildfromdropx86/bin
-    cp ./tests/test.defaults.json buildfromdropx86/bin
-    cp ./tests/test.audio.utterances.json buildfromdropx86/bin
-    cp ./tests/test.subscriptions.regions.json buildfromdropx86/bin
+    [[ ! -d buildfromdropx86/bin/input ]] && cp -rf tests/input buildfromdropx86/bin
+    [[ ! -e buildfromdropx86/bin/test.defaults.json ]] && cp ./tests/test.defaults.json buildfromdropx86/bin
+    [[ ! -e buildfromdropx86/bin/test.audio.utterances.json ]] && cp ./tests/test.audio.utterances.json buildfromdropx86/bin
+    [[ ! -e buildfromdropx86/bin/test.subscriptions.regions.json ]] && cp ./tests/test.subscriptions.regions.json buildfromdropx86/bin
 fi
 
 cmdTimeout=
